refactor: step1에서 사용했던 코드의 패키지 변경
