+++ src/main/java/stringcalculator/domain/Number.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package stringcalculator.domain;
 
 import java.util.Objects;
