+++ src/test/java/stringcalculator/command/CommandFactoryTest.java
@@ -1,4 +1,4 @@
-package lotto.command;
+package stringcalculator.command;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
