+++ src/test/java/stringcalculator/domain/StringArithmeticInputTestTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package stringcalculator.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
